src/requirement.txt
pip
accelerate
diffusers
fastapi
ffmpy
ftfy
huggingface-hub
lit
matplotlib
numpy
onnx
onnxruntime
openai
opencv-python
pandas
pillow
pycocotools
regex
torchaudio
torchvision
tqdm
transformers
scipy
safetensors